+++ app/Views/mhsDetail.php
